\name{declared-internal}

\alias{all_missing_values}
\alias{sort_labelled}
\alias{order_labelled}
\alias{to_labels}
\alias{order_declared}
\alias{sort_declared}
\alias{format_declared}
\alias{is_tagged_string}
\alias{names_values}
\alias{to_labels}

\title{Internal functions}

\description{
  Functions for internal use only.
}

\keyword{internal}
