% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declared_package.R
\docType{package}
\name{declared-package}
\alias{declared-package}
\title{declared: Functions for Declared Missing Values}
\description{
A zero dependency package containing functions to declare labels and missing values, coupled with associated functions to create (weighted) tables of frequencies and various other summary measures. Some of the base functions have been rewritten to make use of the specific information about the missing values, most importantly to distinguish between empty and declared NA values. Some functions have similar functionality with the corresponding ones from packages "haven" and "labelled". The aim is to ensure as much compatibility as possible with these packages, while offering an alternative in the objects of class "declared".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dusadrian.github.io/declared/}
  \item \url{https://github.com/dusadrian/declared}
  \item Report bugs at \url{https://github.com/dusadrian/declared/issues}
}

}
\author{
\strong{Maintainer}: Adrian Dusa \email{dusa.adrian@unibuc.ro} (\href{https://orcid.org/0000-0002-3525-9253}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adrian Dusa \email{dusa.adrian@unibuc.ro} (\href{https://orcid.org/0000-0002-3525-9253}{ORCID}) [copyright holder]
}

}
\keyword{internal}
