% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declared.R, R/internals.R, R/wIQR.R,
%   R/wfivenum.R, R/wmean.R, R/wmedian.R, R/wmode.R, R/wquantile.R, R/wsd.R,
%   R/wstandardize.R, R/wsummary.R, R/wtable.R, R/wvar.R
\name{direct_declared}
\alias{direct_declared}
\alias{declared_internal}
\alias{format_declared}
\alias{order_declared}
\alias{value_labels}
\alias{variable_label}
\alias{names_values}
\alias{makeTag_}
\alias{hasTag_}
\alias{getTag_}
\alias{anyTagged_}
\alias{w_IQR}
\alias{w_fivenum}
\alias{w_mean}
\alias{w_median}
\alias{w_mode}
\alias{w_quantile}
\alias{w_sd}
\alias{w_standardize}
\alias{w_summary}
\alias{w_table}
\alias{w_var}
\title{declared internal functions}
\usage{
direct_declared(
  x,
  na_index = NULL,
  na_values = NULL,
  na_range = NULL,
  labels = NULL,
  label = NULL,
  measurement = NULL,
  decimals = NULL,
  date = inherits(x, "Date")
)

format_declared(x, digits = getOption("digits"))

order_declared(
  x,
  na.last = NA,
  decreasing = FALSE,
  method = c("auto", "shell", "radix"),
  empty.last = na.last,
  ...
)

value_labels(...)

variable_label(...)

names_values(x, drop_na = FALSE, observed = TRUE)

makeTag_(...)

hasTag_(x, tag = NULL)

getTag_(x)

anyTagged_(x)

w_IQR(...)

w_fivenum(...)

w_mean(...)

w_median(...)

w_mode(...)

w_quantile(...)

w_sd(...)

w_standardize(...)

w_summary(...)

w_table(...)

w_var(...)
}
\description{
Functions to be used internally, only by developers and
contributors.
}
\keyword{internal}
