\name{declared-package}

\alias{declared-package}

\docType{package}

\title{
Functions to Declare Missing Values
}

\description{
A set of functions to declare labels and missing values, coupled with associated
functions to create (weighted) tables of frequencies and various other summary
measures.
}

\details{
    \tabular{ll}{
    Package: \tab declared\cr
    Type: \tab Package\cr
    Version: \tab 0.11\cr
    Date: \tab 2022-03-01\cr
    License: \tab GPL-v3\cr
    }
}
\author{
Adrian Dusa

Maintainer: Adrian Dusa (dusa.adrian@unibuc.ro)
}

\keyword{misc}
