\name{declared-package}

\alias{declared-package}

\docType{package}

\title{
Functions to Declare Missing Values
}

\description{
A set of functions that offer an alternative to package 'haven', to deal with
labelled objects with existing values declared as missing.
}

\details{
    \tabular{ll}{
    Package: \tab declared\cr
    Type: \tab Package\cr
    Version: \tab 0.5\cr
    Date: \tab 2021-06-04\cr
    License: \tab GPL-v3\cr
    }
}
\author{
Adrian Dusa

Maintainer: Adrian Dusa (dusa.adrian@unibuc.ro)
}

\keyword{misc}
