\name{Internal functions}

\alias{all_missing_values}
\alias{sort_labelled}
\alias{order_labelled}
\alias{to_labels}
\alias{order_declared}
\alias{sort_declared}
\alias{format_declared}
\alias{is_tagged_string}
\alias{names_values}
\alias{stopError_}
\alias{coerceMode_}
\alias{possibleNumeric_}
\alias{asNumeric_}
\alias{wholeNumeric_}
\alias{tryCatchWEM_}
\alias{padLeft_}
\alias{padRight_}
\alias{padBoth_}
\alias{unlockEnvironment_}
\alias{makeTag_}
\alias{hasTag_}
\alias{getTag_}
\alias{numdec_}
\alias{trimstr_}

\title{Internal functions}

\description{
  Functions for internal use only.
}

\keyword{internal}
