% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{declared_internal}
\alias{declared_internal}
\alias{format_declared}
\alias{order_declared}
\alias{value_labels}
\alias{variable_label}
\alias{names_values}
\title{declared internal functions}
\usage{
format_declared(x, digits = getOption("digits"))

order_declared(
  x,
  na.last = NA,
  decreasing = FALSE,
  method = c("auto", "shell", "radix"),
  empty.last = na.last
)

value_labels(...)

variable_label(...)

names_values(x, drop_na = FALSE)
}
\description{
Functions to be used internally, only by developers and contributors.
}
\keyword{internal}
