% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven.R, R/internals.R
\name{as.haven}
\alias{as.haven}
\alias{declared_internal}
\alias{format_declared}
\alias{order_declared}
\alias{value_labels}
\alias{variable_label}
\title{declared internal functions}
\usage{
as.haven(x, ...)

format_declared(x, digits = getOption("digits"))

order_declared(
  x,
  na.last = NA,
  decreasing = FALSE,
  method = c("auto", "shell", "radix"),
  empty.last = na.last
)

value_labels(...)

variable_label(...)
}
\description{
Only to be used internally for developers and contributors.
}
\keyword{internal}
