\name{About the declared package}

\alias{declared_package}

\docType{package}

\title{
Functions to Declare Missing Values
}

\description{
A set of functions to declare labels and missing values, coupled with
associated functions to create (weighted) tables of frequencies and various
other summary measures. 
Various base functions are rewritten to make use of the specific information
about the missing values, most importantly to distinguish between empty and
declared missing values.
Many functions have a similar functionality with the corresponding functions
from packages "haven" and "labelled". A lot of effort was spent to ensure as
much compatibility as possible with these packages, with the intention to
offer a complementary alternative for the objects of class "declared".
}

\details{
    \tabular{ll}{
    Package: \tab declared\cr
    Type: \tab Package\cr
    Version: \tab 0.15\cr
    Date: \tab 2022-04-13\cr
    License: \tab GPL-v3\cr
    }
}
\author{
Adrian Dusa

Maintainer: Adrian Dusa (dusa.adrian@unibuc.ro)
}

\keyword{misc}
